chore(popup): move not found page
