+++ packages/popup/src/pages/HomePage/index.ts
@@ -1,0 +1,1 @@
+export { default as HomePage } from "./HomePage";
