+++ packages/popup/src/pages/NotFoundPage/index.ts
@@ -1,0 +1,1 @@
+export { default as NotFoundPage } from "./NotFoundPage";
